Add a tailwindUI form for example purposes
I took the first example form (with code provided) from:
https://tailwindui.com/components/application-ui/forms/form-layouts

Tossed it in https://htmltohaml.com/

And put it in as an example of what all the tailwind classes are on the
different forms.

I think it's pretty powerful to be able to have a well thought-out
element to have as an example of how it gets built.

Where I'm going in circles though is:
1. Does seeing how tailwindUI forms get built matter to us when we
   actually already have Hillary thinking about designs?
2. Could we have Hillary maybe using the various tailwindUI examples
   that are laid out as inspiration for how to think about mobile and
   scaling it up from there?