Using Turbo to submit the search form.
I guess the interesting thing here is using that @search variable
as the UI state. Doesn't feel great to have a single variable lurking
around in the controller but this could be organized into a pattern
of the UI state with this different concept of keeping everything
server side.

It's also interesting the turbo idea of "it just works" for form
submission, since it's by default just doing a fetch and re-rendering
with the DOM diff.