Stop using tailwinds stimulus controllers.
We had a stimulus controller defined for the dropdown component,
but importing tailwinds stimulus controllers overwrites stuff and
just makes things go in a different direction that expected with
Stimulus controllers.

When you open a file, you see the controller defined and you expect
that all interactions will be happening inside that related controller
js file.