Don't lose the parent div that we're streaming to when replacing.
Before we were replacing the whole div with our partial and that
made us lose the div where we would keep streaming. Since we have
that parent in the partial too now, when we replace that original
div with the broadcast, we still keep the same dom structure
to keep streaming to.