Validate form+model when saving the model.
When we submitted the form, we were just delegating the save
call to the model and that was validating only the model and
the form wasn't having the model errors to display.

This implements a very naive approach to also have the validation
happening on save in our form just like AR does. Full implementation
should probably included the options to whether you want to skip or not
and investigate more if there's other side effects we want happening in
the form when saving.

CC @shepmaster Let me know what you think about this situation.

Thanks!